Allow user to set the layer offset
