+++ src/hooks/mapUdatingHooks/usePTRoutesLayerUpdate.ts
@@ -1,8 +1,6 @@
-import { AnyLayer, GeoJSONSource, LngLatBounds } from 'mapbox-gl';
-import { useEffect, useRef } from 'react';
+import { LngLatBounds } from 'mapbox-gl';
+import { useEffect } from 'react';
 
-import ptRoutesLayer from '../../data/layers/ptRoutesLayer.json';
-import ptStopsLayer from '../../data/layers/ptStopsLayer.json';
 import { useAppSelector } from '../../store';
 import { routesPaintWhenSelected } from '../useHookUtil';
 
@@ -46,6 +44,7 @@ export const usePTRoutesLayerUpdate = (map: mapboxgl.Map | null): void => {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [selectedPTRouteID]);
 
+    // Apply current filter to routes
     const filteredRoutes = useAppSelector((state) => state.slice.filteredRoutes);
     useEffect(() => {
         if (map && map.getLayer('ptRoutes')) {
@@ -55,62 +54,13 @@ export const usePTRoutesLayerUpdate = (map: mapboxgl.Map | null): void => {
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [filteredRoutes]);
-};
 
-export const layerConfig: { [key: string]: AnyLayer } = {
-    ptRoutes: ptRoutesLayer as AnyLayer,
-    ptStops: ptStopsLayer as AnyLayer,
-};
-// export const useUpdatePTVehiclesLayer = (map: mapboxgl.Map): void => {};
-
-/**
- * Initializes the sources and layers on the map.
- */
-export const useInitializeSourcesAndLayers = (map: mapboxgl.Map | null): React.MutableRefObject<boolean> => {
-    const hasInitialized = useRef(false);
-
+    // Apply the selected route offset to routes
+    const routeOffset = useAppSelector((state) => state.slice.routeOffset);
     useEffect(() => {
-        if (map) {
-            Object.entries(layerConfig).forEach(([layerId, layer]: [string, AnyLayer]) => {
-                const sourceId = `${layerId}Source`;
-
-                if (!map.getSource(sourceId)) {
-                    map.addSource(sourceId, {
-                        type: 'geojson',
-                        data: { type: 'FeatureCollection', features: [] },
-                    });
-
-                    if (!map.getLayer(layer.id)) {
-                        map.addLayer(layer);
-                    }
-                }
-            });
-
-            hasInitialized.current = true;
-            console.log('generate layers');
+        if (map && map.getLayer('ptRoutes')) {
+            map.setPaintProperty('ptRoutes', 'line-offset', routeOffset);
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [map]);
-
-    return hasInitialized;
-};
-
-// Custom local hook for applying data to a source when the data is updated.
-export const useApplyDataToSource = <T extends GeoJSON.Feature>(
-    mapInitialized: boolean,
-    layerId: string,
-    features: T[],
-    map?: mapboxgl.Map | null,
-): void => {
-    useEffect((): void => {
-        if (map && mapInitialized) {
-            const source = map.getSource(`${layerId}Source`) as GeoJSONSource;
-            if (source)
-                source.setData({
-                    type: 'FeatureCollection',
-                    features: features,
-                });
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [mapInitialized, features]);
+    }, [routeOffset]);
 };
